Separate defconst into defaddr and defimmed.
Before, defconst was conflating two different use-cases: addresses that should be output literally into source, and constants that are treated as immediate values to be loaded into A or compared against. Since immediate values need a "#" prefix in 6502 assembly syntax, the parameter to defconst needed to be a string, even though logically the constant was never actually a string. Instead, use defaddr for addresses and defimmed for immediate values.