Symbol-label table, storing constants, addresses, data, vars.
Define built-in values using sym-label table. Output them at the start of the generated code. Determine whether a symbol is an address or constant in order to decide if immediate mode should be used. Merge data-segment into this table as well. Ora instruction performs constant folding in a very hacky way.